Starting off
Include `package.json`
Include extension of `ampersand-input-view`
includes features:
custom `clean` function
`addTest` function.
`tests` are now a `props` so changes in the `tests` array will fire
events
`maxlength` with `maxlengthExceededMessage` informs the user that they
nee shorter input. Use `:num` in the message to be replaced by the
current `maxlength`
function `disable` to set `disabled = true` which sets the attribute on
the `input` element
`attrs` is a child `state` that allows extra properties. When change
events fire the `attrs` properties get added as attributes on the
element
Fixed a bug exposed in IE. IE kills execution when an input type is
change to an unsupported type ('month', 'date', etc.) so I swallow the
exception and fall back on the default `type=`text`